requirements.txt
djangorestframework
djangorestframework-simplejwt==5.0.0
django==3.2
psycopg2-binary==2.8.4
requests